+++ energyquantified/api/__init__.py
@@ -4,6 +4,7 @@ from .timeseries import TimeseriesAPI
 from .periods import PeriodsAPI
 from .period_instances import PeriodInstancesAPI
 from .ohlc import OhlcAPI
+from .srmc import SrmcAPI
 
 __all__ = [
     "InstancesAPI",
@@ -12,4 +13,5 @@ __all__ = [
     "PeriodsAPI",
     "PeriodInstancesAPI",
     "OhlcAPI",
+    "SrmcAPI",
 ]